KVM: x86: Move user return msr code into VAC
